fix bug related to network simulation + minor changes
Every operation done by the robot opened and closed a TCP socket. The close operation was graceful and so happened only when the transmission buffer was empty.
In case of operation interrupted for maximum duration (e.g. transfer for throughput measurement), the transfer continued in background.

ns-3 does not support the SO_LINGER option of UNIX sockets.
However, I implemented a simplified version of the option which closes immediately the connection, without sending the bytes left in the transmission buffer.