update mitigation rules
I removed the rotation forcing in the V-REP script, because in some situation the navigation tried to rotate in the other sense and the robot got stuck.
To fix the issue of robot stuck in front of an obstacle, I modified some rules. Now there are no rules that stop the robot completely.